chore: 이벤트 주석 생성 및 메시지 이넘 타입 이름 변경
