+++ src/main/java/com/sprint/ootd5team/domain/notification/event/type/ClothesAttributeUpdateEvent.java
@@ -5,6 +5,12 @@ import com.sprint.ootd5team.domain.notification.enums.NotificationTemplateType;
 import java.util.List;
 import java.util.UUID;
 
+/**
+ * CLOTHES_ATTRIBUTE_UPDATE("의상 속성이 변경되었어요.", "[%s]속성을 확인해보세요.");
+ * 속성 이름
+ *
+ * receiverId = 모든 유저 id
+ */
 public class ClothesAttributeUpdateEvent extends DomainEvent<ClothesAttributeDefDto> {
 
     public ClothesAttributeUpdateEvent(ClothesAttributeDefDto data) {
@@ -13,7 +19,7 @@ public class ClothesAttributeUpdateEvent extends DomainEvent<ClothesAttributeDef
 
     @Override
     public NotificationTemplateType getTemplateType() {
-        return NotificationTemplateType.CLOTHES_ATTRIBUTE_UPDATE;
+        return NotificationTemplateType.CLOTHES_ATTRIBUTE_UPDATED;
     }
 
     @Override
