+++ src/main/java/com/sprint/ootd5team/domain/notification/event/type/DmCreateEvent.java
@@ -4,6 +4,12 @@
 //import java.util.List;
 //import java.util.UUID;
 //
+///**
+// * DM_RECEIVED("[DM] %s", "%s"),
+// * 메시지 발신자(sender) 이름, 메시지 내용
+// *
+// * receiverId = 수신자 id
+// */
 //public class DmCreateEvent extends DomainEvent<DirectMessageDto>{
 //    public DmCreateEvent(DirectMessageDto data) {
 //        super(data);
