+++ src/main/java/com/sprint/ootd5team/domain/notification/event/type/FollowCreateEvent.java
@@ -4,6 +4,12 @@
 //import java.util.List;
 //import java.util.UUID;
 //
+///**
+// * FOLLOWED("%님이 나를 팔로우 했어요", "%s"),
+// * 팔로우를 요청한 사람 이름(follower)
+// *
+// * receiverId = 팔로우를 요청 받은 사람 이름(followee)
+// */
 //public class FollowCreateEvent extends DomainEvent<FollowDto>{
 //
 //    public FollowCreateEvent(FollowDto data) {
